+++ R/functions-mgf.R
@@ -13,6 +13,8 @@
 ##'
 ##' @importFrom MsCoreUtils rbindFill
 ##'
+##' @importFrom methods as
+##'
 ##' @author Laurent Gatto, Johannes Rainer
 ##'
 ##' @noRd
@@ -107,4 +109,57 @@
     res$mz = list(ms[, 1L])
     res$intensity = list(ms[, 2L])
     res
+}
+
+#' @description
+#'
+#' Function to export a `Spectra` object in MGF format to `con`.
+#'
+#' @param x `Spectra`
+#'
+#' @param con output file.
+#'
+#' @param mapping named `character` vector that maps from `spectraVariables`
+#'    (i.e. `names(mapping)`) to the variable name that should be used in the
+#'    MGF file.
+#'
+#' @author Johannes Rainer
+#'
+#' @importMethodsFrom Spectra spectraVariables spectraNames as.list spectraData
+#'
+#' @noRd
+#'
+#' @examples
+#'
+#' spd <- DataFrame(msLevel = c(2L, 2L, 2L), rtime = c(1, 2, 3))
+#' spd$mz <- list(c(12, 14, 45, 56), c(14.1, 34, 56.1), c(12.1, 14.15, 34.1))
+#' spd$intensity <- list(c(10, 20, 30, 40), c(11, 21, 31), c(12, 22, 32))
+#'
+#' sps <- Spectra(spd)
+#'
+#' .export_mgf(sps)
+.export_mgf <- function(x, con = stdout(), mapping = spectraVariableMapping()) {
+    spv <- spectraVariables(x)
+    spd <- spectraData(x, spv[!(spv %in% c("dataOrigin", "dataStorage"))])
+    idx <- match(colnames(spd), names(mapping))
+    colnames(spd)[!is.na(idx)] <- mapping[idx[!is.na(idx)]]
+    l <- nrow(spd)
+    tmp <- lapply(colnames(spd), function(z) {
+        paste0(z, "=", spd[, z], "\n")
+    })
+    if (!any(colnames(spd) == "TITLE")) {
+        if (!is.null(spectraNames(x)))
+            title <- paste0("TITLE=", spectraNames(x), "\n")
+        else
+            title <- paste0("TITLE=msLevel ", spd$msLevel, "; retentionTime ",
+                            spd$rtime, "; scanNum ", spd$acquisitionNum, "\n")
+        tmp <- c(list(title), tmp)
+    }
+    pks <- vapply(as.list(x), function(z)
+        paste0(paste0(z[, 1], " ", z[, 2], "\n"), collapse = ""),
+        character(1))
+    tmp <- do.call(cbind, c(list(rep_len("BEGIN IONS\n", l)),
+                            tmp, list(pks), list(rep_len("END IONS\n", l))))
+    tmp[grep("=NA\n", tmp)] <- ""
+    writeLines(apply(tmp, 1, paste0, collapse = ""), con = con)
 }